Added Info Window, Sorting & Focus on Target
1. Sorting of info-window based on sort-field -> Date, String & Number supported
2. Focus on map target either from clicking the point or selecting  the item on RecyclerView
3. Render patient information on Info Window

+++ library/src/main/java/io/ona/kujaku/activities/MapActivity.java
@@ -1,15 +1,20 @@
 package io.ona.kujaku.activities;
 
+import android.app.Activity;
 import android.graphics.PointF;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.util.DisplayMetrics;
 import android.util.Log;
 import android.view.View;
 
-import com.google.gson.JsonElement;
 import com.mapbox.mapboxsdk.Mapbox;
+import com.mapbox.mapboxsdk.camera.CameraPosition;
+import com.mapbox.mapboxsdk.camera.CameraUpdateFactory;
 import com.mapbox.mapboxsdk.geometry.LatLng;
 import com.mapbox.mapboxsdk.maps.MapView;
 import com.mapbox.mapboxsdk.maps.MapboxMap;
@@ -21,18 +26,19 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 
 import io.ona.kujaku.R;
 import io.ona.kujaku.adapters.InfoWindowAdapter;
 import io.ona.kujaku.adapters.InfoWindowObject;
 import io.ona.kujaku.helpers.MapBoxStyleStorage;
+import io.ona.kujaku.sorting.Sorter;
 import io.ona.kujaku.sorting.objects.SortField;
 import io.ona.kujaku.utils.Permissions;
+import io.ona.kujaku.views.InfoWindowLayoutManager;
 import utils.Constants;
+import utils.helpers.converters.GeoJSONFeature;
 
 /**
  * This activity displays a MapView once provided with a a MapBox Access Key & String array.
@@ -57,13 +63,23 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
     private JSONObject mapboxStyleJSON;
     private static final String TAG = MapActivity.class.getSimpleName();
 
-    private LinkedHashMap<String, InfoWindowObject> featureList = new LinkedHashMap<>();
+    private LinkedHashMap<String, InfoWindowObject> featuresMap = new LinkedHashMap<>();
+    private ArrayList featureIdList = new ArrayList();
     private MapboxMap mapboxMap;
     private boolean infoWindowDisplayed = false;
 
+    private RecyclerView.OnScrollListener onScrollListener = null;
+
     // Info window stuff
     private RecyclerView infoWindowsRecyclerView;
+    private InfoWindowLayoutManager linearLayoutManager;
     private int lastSelected = 0;
+
+    private int animateToNewTargetDuration = 1000;
+    private int screenWidth = 0;
+
+    private InfoWindowAdapter infoWindowAdapter;
+    private boolean startedScrolling = false;
 
     //Todo: Move reading data to another Thread
 
@@ -74,6 +90,7 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
         initializeViews();
 
         checkPermissions();
+        screenWidth = getScreenWidth(this);
 
         if (getIntent() != null) {
             Bundle bundle = getIntent().getExtras();
@@ -91,10 +108,11 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
 
                         // Extract kujaku meta-data
                         try {
-                            dataLayers = extractSourceNames(mapboxStyleJSON);
                             sortFields = extractSortFields(mapboxStyleJSON);
-                            featureList = extractLayerData(mapboxStyleJSON, dataLayers);
-                            displayInitialFeatures(featureList);
+                            dataLayers = extractSourceNames(mapboxStyleJSON, sortFields);
+                            featuresMap = extractLayerData(mapboxStyleJSON, dataLayers);
+                            featuresMap = sortData(featuresMap, sortFields);
+                            displayInitialFeatures(featuresMap);
                         } catch (JSONException e) {
                             Log.e(TAG, Log.getStackTraceString(e));
                         }
@@ -130,7 +148,7 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
         infoWindowsRecyclerView = (RecyclerView) findViewById(R.id.rv_mapActivity_infoWindow);
     }
 
-    private String[] extractSourceNames(@NonNull JSONObject jsonObject) throws JSONException {
+    private String[] extractSourceNames(@NonNull JSONObject jsonObject, @NonNull SortField[] sortFields) throws JSONException {
         if (jsonObject.has("metadata")) {
             JSONObject metadata = jsonObject.getJSONObject("metadata");
             if (metadata.has("kujaku")) {
@@ -139,7 +157,7 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
                     JSONArray jsonArray = kujakuRelatedData.getJSONArray("data_source_names");
                     String[] dataLayers = new String[jsonArray.length()];
 
-                    for(int i = 0; i < sortFields.length; i++) {
+                    for(int i = 0; i < dataLayers.length; i++) {
                         dataLayers[i] = jsonArray.getString(i);
                     }
 
@@ -154,22 +172,33 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
     private LinkedHashMap<String, InfoWindowObject> extractLayerData(@NonNull JSONObject mapBoxStyleJSON, @NonNull String[] dataSourceNames) throws JSONException {
         if (mapBoxStyleJSON.has("sources")) {
             JSONObject sources = mapBoxStyleJSON.getJSONObject("sources");
-            LinkedHashMap<String, InfoWindowObject> featureList = new LinkedHashMap<>();
+            LinkedHashMap<String, InfoWindowObject> featuresMap = new LinkedHashMap<>();
             int counter = 0;
 
             for(String dataSourceName: dataSourceNames) {
                 if (sources.has(dataSourceName)) {
                     JSONObject jsonObject = sources.getJSONObject(dataSourceName);
-                    if (jsonObject.has("id")) {
-                        String id = jsonObject.getString("id");
-                        //Todo: Should check for errors here & print them in LogCat or notify the dev somehow
-                        featureList.put(id, new InfoWindowObject(counter, jsonObject));
-                        counter++;
+                    if (jsonObject.has("data")) {
+                        JSONObject sourceDataJSONObject = jsonObject.getJSONObject("data");
+                        if (sourceDataJSONObject.has("features")) {
+                            JSONArray featuresJSONArray = sourceDataJSONObject.getJSONArray("features");
+                            for(int i = 0; i < featuresJSONArray.length(); i++) {
+                                JSONObject featureJSON = featuresJSONArray.getJSONObject(i);
+
+                                String id = getFeatureId(featureJSON);
+                                if (!id.isEmpty()) {
+                                    //Todo: Should check for errors here & print them in LogCat or notify the dev somehow
+                                    featuresMap.put(id, new InfoWindowObject(counter, featureJSON));
+                                    featureIdList.add(id);
+                                    counter++;
+                                }
+                            }
+                        }
                     }
                 }
             }
 
-            return featureList;
+            return featuresMap;
         }
 
         return null;
@@ -198,7 +227,7 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
 
     private JSONObject getStyleJSON(@NonNull String stylePathOrJSON) {
         try {
-            return new JSONObject(stylePathOrJSON);
+            return new JSONObject(getStyleJSONString(stylePathOrJSON));
         } catch (JSONException e) {
             Log.e(TAG, Log.getStackTraceString(e));
             return new JSONObject();
@@ -220,24 +249,48 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
         }
     }
 
-    private void displayInitialFeatures(@NonNull LinkedHashMap<String, InfoWindowObject> featureList) {
-        if (!featureList.isEmpty()) {
-            InfoWindowAdapter infoWindowAdapter = new InfoWindowAdapter(featureList, infoWindowsRecyclerView);
+    private void displayInitialFeatures(@NonNull LinkedHashMap<String, InfoWindowObject> featuresMap) {
+        if (!featuresMap.isEmpty()) {
+            infoWindowAdapter = new InfoWindowAdapter(this, featuresMap, infoWindowsRecyclerView);
+            infoWindowAdapter.setOnClickListener(new InfoWindowAdapter.OnClickListener() {
+                @Override
+                public void onClick(View v, int position) {
+                    focusOnFeature(position);
+                }
+            });
+
+            infoWindowsRecyclerView.setHasFixedSize(true);
+            linearLayoutManager = new InfoWindowLayoutManager(this, LinearLayoutManager.HORIZONTAL, false);
+            infoWindowsRecyclerView.setLayoutManager(linearLayoutManager);
             infoWindowsRecyclerView.setAdapter(infoWindowAdapter);
         }
     }
 
-    private LinkedHashMap<String, InfoWindowObject> sortData(@NonNull LinkedHashMap<String, InfoWindowObject> featureList,@NonNull SortField[] sortFields) {
+    private LinkedHashMap<String, InfoWindowObject> sortData(@NonNull LinkedHashMap<String, InfoWindowObject> featuresMap,@NonNull SortField[] sortFields) throws JSONException {
         //TODO: Add support for multiple sorts
+        int counter = 0;
         if (sortFields.length > 0) {
             SortField sortField = sortFields[0];
             if (sortField.getType() == SortField.FieldType.DATE) {
                 //Todo: Add sorter here
-                return featureList;
+                //Todo: Change the order of ids' in the featureIdsList
+                Sorter sorter = new Sorter(new ArrayList(featuresMap.values()));
+                ArrayList<InfoWindowObject> infoWindowObjectArrayList = sorter.mergeSort(0, featuresMap.size() -1, sortField.getDataField(), sortField.getType());
+
+                featuresMap.clear();
+
+                for(InfoWindowObject infoWindowObject: infoWindowObjectArrayList) {
+                    String id = getFeatureId(infoWindowObject);
+                    if (!id.isEmpty()) {
+                        infoWindowObject.setPosition(counter);
+                        featuresMap.put(id, infoWindowObject);
+                        counter++;
+                    }
+                }
             }
         }
 
-        return featureList;
+        return featuresMap;
     }
 
     @Override
@@ -268,7 +321,7 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
     protected void onDestroy() {
         super.onDestroy();
 
-        if (currentStylePath != null && currentStylePath.startsWith("file://")) {
+        if (currentStylePath != null && currentStylePath.startsWith("file://") && currentStylePath.contains(MapBoxStyleStorage.DIRECTORY)) {
             new MapBoxStyleStorage()
                     .deleteFile(currentStylePath.replace("file://", ""), true);
         }
@@ -310,23 +363,16 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
 
         // Get the first feature within the list if one exist
         if (features.size() > 0) {
-            com.mapbox.services.commons.geojson.Feature feature = features.get(0);
+            for(Feature feature: features) {
 
-            // Ensure the feature has properties defined
-            if (feature.getProperties() != null) {
-                if (feature.hasProperty("id")) {
-                    String id =  feature.getId();
-                    if (featureList.containsKey(id)) {
-                        showInfoWindow();
-                        centerMap(point);
-                        renderInfoWindow(featureList.get(id).getPosition());
+                // Ensure the feature has properties defined
+                if (feature.getProperties() != null) {
+                    if (feature.hasProperty("id")) {
+                        String id = feature.getProperty("id").getAsString();
+                        if (featuresMap.containsKey(id)) {
+                            focusOnFeature(id);
+                        }
                     }
-                }
-
-                for (Map.Entry<String, JsonElement> entry : feature.getProperties().entrySet()) {
-                    // Log all the properties
-
-                    Log.d(TAG, String.format("%s = %s", entry.getKey(), entry.getValue()));
                 }
             }
         }
@@ -337,16 +383,184 @@ public class MapActivity extends AppCompatActivity implements MapboxMap.OnMapCli
         if (!infoWindowDisplayed) {
             // Good enough for now
             infoWindowsRecyclerView.setVisibility(View.VISIBLE);
+            infoWindowDisplayed = true;
         }
     }
 
-    private void renderInfoWindow(int position) {
+    private void renderInfoWindow(final int position) {
         if (position > 1) {
 
+            infoWindowAdapter.focusOnPosition(position);
+
+            // Supposed to scroll to the selected position
+            final InfoWindowAdapter.InfoWindowViewHolder infoWindowViewHolder = (InfoWindowAdapter.InfoWindowViewHolder) infoWindowsRecyclerView.findViewHolderForAdapterPosition(position);
+            if (infoWindowViewHolder == null) {
+
+                linearLayoutManager.setOnScrollListener(new InfoWindowLayoutManager.OnScrollListener() {
+                    @Override
+                    public void onFinishedScrolling() {
+                        InfoWindowAdapter.InfoWindowViewHolder infoWindowViewHolder = (InfoWindowAdapter.InfoWindowViewHolder) infoWindowsRecyclerView.findViewHolderForAdapterPosition(position);
+                        if (infoWindowViewHolder != null) {
+
+                            View v = infoWindowViewHolder.itemView;
+                            infoWindowViewHolder.select();
+                            linearLayoutManager.removeOnScrollListener();
+                        }
+                    }
+                });
+                infoWindowsRecyclerView.smoothScrollToPosition(position);
+
+            } else {
+                View v = infoWindowViewHolder.itemView;
+                infoWindowViewHolder.select();
+                animateToPosition(v, position, animateToNewTargetDuration);
+            }
         }
     }
 
     private void centerMap(@NonNull LatLng point) {
-        //Todo: finish this
+        CameraPosition cameraPosition = new CameraPosition.Builder()
+                .target(point)
+                .build();
+
+
+        if (mapboxMap != null) {
+            mapboxMap.animateCamera(CameraUpdateFactory.newCameraPosition(cameraPosition), animateToNewTargetDuration);
+        }
+    }
+
+
+    private void animateToPosition(@NonNull View view, final int position, final int animationDuration) {
+        final int left = view.getLeft()
+                , right = view.getRight();
+        
+        int scrollX = (left - (screenWidth/2)) + (view.getWidth()/2);
+        final int offset = (screenWidth/2) - (view.getWidth()/2);
+
+        infoWindowsRecyclerView.smoothScrollToPosition(position);
+        //Todo Figure out how to handle another item being selected while this one is being animated
+        /*new Thread(new Runnable() {
+            @Override
+            public void run() {
+                int sleepCounts = 10;
+                int sleepDuration = animationDuration/sleepCounts;
+                int counter = 0;
+                int eachMove = (offset - left)/sleepCounts;
+                int startOffset = left;
+
+                while (counter < sleepCounts) {
+                    startOffset += eachMove;
+                    counter++;
+
+                    try {
+                        Thread.sleep(sleepDuration);
+
+                        final int finalStartOffset = startOffset;
+                        runOnUiThread(new Runnable() {
+                            @Override
+                            public void run() {
+                                linearLayoutManager.scrollToPositionWithOffset(position, finalStartOffset);
+                                //linearLayoutManager.smoothScrollToPosition();
+
+                                //Todo add size animation
+                            }
+                        });
+                    } catch (InterruptedException e) {
+                        Log.e(TAG, Log.getStackTraceString(e));
+                    }
+                }
+            }
+        }).start(); */
+    }
+
+    private int getScreenWidth(Activity activity) {
+        // Get the screen width
+        DisplayMetrics displayMetrics = new DisplayMetrics();
+        activity.getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
+        return displayMetrics.widthPixels;
+    }
+
+    private void focusOnFeature(int position, @Nullable String id, @Nullable LatLng latLng) {
+        if (position == -1 && id != null && !id.isEmpty()) {
+            if (featuresMap.containsKey(id)) {
+                position = featuresMap.get(id).getPosition();
+            } else {
+                return;
+            }
+        }
+
+        if (id == null && position > -1) {
+            if (featureIdList.isEmpty()) {
+                return;
+            }
+
+            id = (String) featureIdList.get(position);
+        }
+
+
+        if (latLng == null) {
+            latLng = getFeaturePoint(featuresMap.get(id)
+                    .getJsonObject());
+
+            if (latLng == null) {
+                return;
+            }
+        }
+
+        showInfoWindow();
+        centerMap(latLng);
+        renderInfoWindow(position);
+    }
+
+
+    private void focusOnFeature(int position) {
+        focusOnFeature(position, null, null);
+    }
+
+    private void focusOnFeature(String id, LatLng latLng) {
+        focusOnFeature(-1, id, latLng);
+    }
+
+    private void focusOnFeature(String id) {
+        focusOnFeature(-1, id, null);
+    }
+
+    private LatLng getFeaturePoint(JSONObject featureJSON) {
+        if (featureJSON.has("geometry")) {
+            try {
+                JSONObject featureGeometry = featureJSON.getJSONObject("geometry");
+                if (featureGeometry.has("type") && GeoJSONFeature.Type.POINT.toString().equalsIgnoreCase(featureGeometry.getString("type"))) {
+                    if (featureGeometry.has("coordinates")) {
+                        JSONArray coordinatesArray = featureGeometry.getJSONArray("coordinates");
+                        if (coordinatesArray.length() > 2) {
+                            double lng = coordinatesArray.getDouble(0);
+                            double lat = coordinatesArray.getDouble(1);
+
+                            return (new LatLng(lat, lng));
+                        }
+                    }
+                }
+            } catch (JSONException e) {
+                Log.e(TAG, Log.getStackTraceString(e));
+            }
+        }
+
+        return null;
+    }
+
+    private String getFeatureId(@NonNull InfoWindowObject infoWindowObject) throws JSONException {
+        JSONObject jsonObject = infoWindowObject.getJsonObject();
+        return getFeatureId(jsonObject);
+    }
+
+    private String getFeatureId(@NonNull JSONObject jsonObject) throws JSONException {
+        if (jsonObject.has("properties")) {
+            JSONObject propertiesJSON = jsonObject.getJSONObject("properties");
+            if (propertiesJSON.has("id")) {
+                return propertiesJSON.getString("id");
+            }
+        }
+
+        return "";
     }
 }
